+++ classes/output/courseformat/content/section/controlmenu.php
@@ -140,8 +140,8 @@ class controlmenu extends \core_courseformat\output\local\content\section\contro
         }
 
         if (
-            $section->parent && has_capability('moodle/course:update', $coursecontext) &&
-                $section->section != $this->format->get_viewed_section()
+            $section->section != $this->format->get_viewed_section() &&
+                $this->format->can_mergeup_section($section)
         ) {
             $mergeupurl = new moodle_url($url, ['mergeup' => $section->section]);
             $controls['mergeup'] = new \core\output\action_menu\link_secondary(